rename ilink-purpose
