+++ src/common/types/property-bag/link-purpose-property-bag.ts
